@@ -24,23 +24,23 @@
   "license": "ISC",
   "homepage": "https://lern.com.br",
   "dependencies": {
-    "@babel/core": "^7.0.0-beta.36",
-    "@babel/plugin-transform-react-jsx": "^7.0.0-beta.36",
-    "@babel/runtime": "^7.0.0-beta.36",
-    "@material-ui/core": "^1.0.0",
+    "@babel/core": "^7.0.0-beta.47",
+    "@babel/plugin-transform-react-jsx": "^7.0.0-beta.47",
+    "@babel/runtime": "^7.0.0-beta.47",
     "autosuggest-highlight": "^3.1.1",
-    "aws-sdk": "^2.198.0",
-    "chart.js": "^2.7.1",
-    "draft-js": "^0.10.4",
-    "lodash": "^4.17.4",
+    "aws-sdk": "^2.245.1",
+    "chart.js": "^2.7.2",
+    "draft-js": "^0.10.5",
+    "lodash": "^4.17.10",
     "loglevel": "^1.6.1",
+    "material-ui": "^1.0.0-beta.25",
     "material-ui-icons": "^1.0.0-beta.36",
     "meteor-node-stubs": "^0.3.2",
-    "moment": "^2.19.2",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
+    "moment": "^2.22.1",
+    "prop-types": "^15.6.1",
+    "react": "^16.4.0",
     "react-autosuggest": "^9.3.4",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.4.0",
     "react-mounter": "^1.2.0",
     "react-tap-event-plugin": "^3.0.2",
     "sudoku": "0.0.2"
