@@ -18,55 +18,55 @@ import Safe from './components/Safe.jsx';
 const muiTheme = createMuiTheme({
   palette: {
     primary: {
-      '50': '#e3f2fd',
-      '100': '#bbdefb',
-      '200': '#90caf9',
-      '300': '#64b5f6',
-      '400': '#42a5f5',
-      '500': '#2196f3',
-      '600': '#1e88e5',
-      '700': '#1976d2',
-      '800': '#1565c0',
-      '900': '#0d47a1',
-      'A100': '#82b1ff',
-      'A200': '#448aff',
-      'A400': '#2979ff',
-      'A700': '#2962ff',
-      'contrastDefaultColor': 'light'
+      50: '#e3f2fd',
+      100: '#bbdefb',
+      200: '#90caf9',
+      300: '#64b5f6',
+      400: '#42a5f5',
+      500: '#2196f3',
+      600: '#1e88e5',
+      700: '#1976d2',
+      800: '#1565c0',
+      900: '#0d47a1',
+      A100: '#82b1ff',
+      A200: '#448aff',
+      A400: '#2979ff',
+      A700: '#2962ff',
+      contrastDefaultColor: 'light',
     },
-    'secondary': {
-      '50': '#fff8e1',
-      '100': '#ffecb3',
-      '200': '#ffe082',
-      '300': '#ffd54f',
-      '400': '#ffca28',
-      '500': '#ffc107',
-      '600': '#ffb300',
-      '700': '#ffa000',
-      '800': '#ff8f00',
-      '900': '#ff6f00',
-      'A100': '#ffe57f',
-      'A200': '#ffd740',
-      'A400': '#ffc400',
-      'A700': '#ffab00',
-      'contrastDefaultColor': 'light'
+    secondary: {
+      50: '#fff8e1',
+      100: '#ffecb3',
+      200: '#ffe082',
+      300: '#ffd54f',
+      400: '#ffca28',
+      500: '#ffc107',
+      600: '#ffb300',
+      700: '#ffa000',
+      800: '#ff8f00',
+      900: '#ff6f00',
+      A100: '#ffe57f',
+      A200: '#ffd740',
+      A400: '#ffc400',
+      A700: '#ffab00',
+      contrastDefaultColor: 'light',
     },
-    'error': {
-      '50': '#ffebee',
-      '100': '#ffcdd2',
-      '200': '#ef9a9a',
-      '300': '#e57373',
-      '400': '#ef5350',
-      '500': '#f44336',
-      '600': '#e53935',
-      '700': '#d32f2f',
-      '800': '#c62828',
-      '900': '#b71c1c',
-      'A100': '#ff8a80',
-      'A200': '#ff5252',
-      'A400': '#ff1744',
-      'A700': '#d50000',
-      'contrastDefaultColor': 'light'
+    error: {
+      50: '#ffebee',
+      100: '#ffcdd2',
+      200: '#ef9a9a',
+      300: '#e57373',
+      400: '#ef5350',
+      500: '#f44336',
+      600: '#e53935',
+      700: '#d32f2f',
+      800: '#c62828',
+      900: '#b71c1c',
+      A100: '#ff8a80',
+      A200: '#ff5252',
+      A400: '#ff1744',
+      A700: '#d50000',
+      contrastDefaultColor: 'light',
     },
   },
 });
@@ -90,24 +90,24 @@ class LayoutView extends React.Component {
 
   constructor(props) {
     super(props);
+    i18n.onChangeLocale(this.onLocale.bind(this));
     this.state = { locale: 'en-US' };
   }
 
-  componentWillReceiveProps(props) {
-    const { user } = props;
+  static getDerivedStateFromProps(nextProps, prevState) {
+    const { user } = nextProps;
+    if (_.get(user, 'profile.locale')) {
+      if (user.profile.locale != prevState.locale) {
+        i18n.setLocale(_.get(user, 'profile.locale'));
+        return ({ locale: _.get(user, 'profile.locale') });
+      }
+    }
 
-    if (_.get(user, 'profile.locale'))
-      i18n.setLocale(_.get(user, 'profile.locale'));
-
-    this.setState({ locale: _.get(user, 'profile.locale') });
+    return null;
   }
 
   onLocale(locale) {
     this.setState({ locale });
-  }
-
-  componentWillMount() {
-    i18n.onChangeLocale(this.onLocale.bind(this));
   }
 
   componentWillUnmount() {
