@@ -11,11 +11,26 @@ import _ from 'lodash';
  */
 class Safe extends React.Component {
 
-  /* Methods
+  /* Lifecycle
   */
 
-  updateAccess({ protect, user, logging }) {
-    this.setState({
+  constructor(props) {
+    super(props);
+    const { protect, user, logging } = props;
+    this.state = {
+      access: (
+         !protect ? true
+       : logging ? undefined
+       : !user ? null
+       : _.isEmpty(user.roles) ? undefined
+       : user.hasRole(protect)
+     ),
+    };
+  }
+
+  static getDerivedStateFromProps(nextProps, prevState) {
+    const { protect, user, logging } = nextProps;
+    return ({
       access: (
          !protect ? true
        : logging ? undefined
@@ -26,25 +41,7 @@ class Safe extends React.Component {
     });
   }
 
-  /* Lifecycle
-  */
-
-  constructor(props) {
-    super(props);
-    this.state = { access: undefined };
-  }
-
-  componentWillMount() {
-    const updateAccess = this.updateAccess.bind(this);
-    updateAccess(this.props);
-  }
-
-  componentWillReceiveProps(props) {
-    const updateAccess = this.updateAccess.bind(this);
-    updateAccess(props);
-  }
-
-  componentWillUpdate({ user }, { access }) {
+  componentDidUpdate({ user }, { access }) {
     if (access === null) {
       snack('Você deve entrar primeiro');
       FlowRouter.go('PublicLogin');
